extract PdfSinglePageRenderer
